Move editors logic into diagram package
