+++ gaphor/diagram/general/__init__.py
@@ -1,0 +1,5 @@
+from .comment import CommentItem
+from .commentline import CommentLineItem
+from .simpleitem import Box, Ellipse, Line
+
+from . import connectors
